Fix skills circular ref
